requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-core
langchain_ollama
sentence-transformers
faiss-cpu
pypdf
huggingface-hub
transformers
torch
requests
numpy
